fix problem mit double to int convert

+++ wetter/Models/WettwerResponses/CurrentWeatherResponse/CurrentWeather.cs
@@ -54,6 +54,6 @@ namespace wetter.Models.WettwerResponses.CurrentWeatherResponse
         public double WindGusts { get; set; }
 
         [JsonPropertyName("is_day")]
-        public int IsDay { get; set; }
+        public double IsDay { get; set; }
     }
 }
